Refactor -> Rename went too far
