+++ fe1-web/features/rollCall/screens/RollCallOpened.tsx
@@ -41,7 +41,8 @@ const tokenMatcher = new RegExp('^[A-Za-z0-9_-]{43}=$');
 const RollCallOpened = () => {
   const route = useRoute();
   const { rollCallID, time } = route.params;
-  const navigation = useNavigation();
+  // FIXME: Navigation should use a defined type here (instead of any)
+  const navigation = useNavigation<any>();
   const [attendees, updateAttendees] = useState(new Set<string>());
   const [inputModalIsVisible, setInputModalIsVisible] = useState(false);
   const toast = useToast();
@@ -100,7 +101,6 @@ const RollCallOpened = () => {
     const attendeesList = Array.from(attendees).map((key: string) => new PublicKey(key));
 
     return requestCloseRollCall(updateId, attendeesList).then(() => {
-      // @ts-ignore
       navigation.navigate(STRINGS.organizer_navigation_tab_home);
     }).catch((err) => {
       toast.show(`Could not close roll call, error: ${err}`, {
