+++ fe1-web/features/social/screens/SocialProfile.tsx
@@ -13,7 +13,7 @@ import STRINGS from 'res/strings';
 import socialMediaProfileStyles from '../styles/socialMediaProfileStyles';
 import { ChirpCard } from '../components';
 import { Chirp, ChirpState } from '../objects';
-import { makeChirpsListOfUser } from '../reducer/SocialReducer';
+import { makeChirpsListOfUser } from '../reducer';
 
 /**
  * UI for the profile of the current user.
