+++ fe1-web/features/social/screens/SocialUserProfile.tsx
@@ -12,7 +12,7 @@ import STRINGS from 'res/strings';
 import { ChirpCard } from '../components';
 import socialMediaProfileStyles from '../styles/socialMediaProfileStyles';
 import { Chirp, ChirpState } from '../objects';
-import { makeChirpsListOfUser } from '../reducer/SocialReducer';
+import { makeChirpsListOfUser } from '../reducer';
 
 /**
  * UI for the profile of a user.
